Added a "Type" metadata field which can be "light" or "dark".
I may later remove the "Dark" metadata field, and use the "Type" field and the suffix on the filename to indentify if a spectrum is light or dark.

A light spectrum is one recorded when light is entering the spectrometer (or spectrometers). A dark spectrum is acquired when no light is entering the spectrometer(s).